fix broken spop test
